Create `tags` files for use in Concourse
The plan is to build the CSW related images in Concourse, but they need
to be tagged as both "latest" and with a version number. This is
possible using the `additional_tags` parameter in the registry-image
resource output step (see
https://github.com/concourse/registry-image-resource#parameters-1).

These files should have their version bumped when a change is made to
the Dockerfiles. I shall be updating the README to reflect that fact.